pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "openai>=2.31.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
]
